+++ src/data/print_documents/managers.py
@@ -107,6 +107,7 @@ class PrintDocumentOptionManager(EventBoundEntityManager[PrintOption]):
             options.IndividualTeamMinGenderCountPrintOption,
             options.IndividualTeamMaxPerEntityPrintOption,
             options.IndividualTeamDisplayIncompletePrintOption,
+            options.FixedBoardOrderPrintOption,
         ]
         plugin_manager.hook_for_event(self.event, 'insert_print_option')(
             print_options=print_options